Moved the resize util into event-util
